+++ src/datasets/mod.rs
@@ -1,4 +1,5 @@
 mod as2org;
+mod as2rel;
 mod country;
 mod ip;
 mod pfx2as;
@@ -6,6 +7,7 @@ mod radar;
 mod rpki;
 
 pub use crate::datasets::as2org::*;
+pub use crate::datasets::as2rel::*;
 pub use crate::datasets::country::*;
 pub use crate::datasets::ip::*;
 pub use crate::datasets::pfx2as::*;